+++ engine/maths/nmatrix.h
@@ -67,9 +67,9 @@ namespace regina {
 template <class T>
 class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
     protected:
-        size_t nRows;
+        unsigned long nRows;
             /**< The number of rows in the matrix. */
-        size_t nCols;
+        unsigned long nCols;
             /**< The number of columns in the matrix. */
         T** data;
             /**< The actual entries in the matrix.
@@ -88,9 +88,9 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          * @param rows the number of rows in the new matrix.
          * @param cols the number of columns in the new matrix.
          */
-        NMatrix(size_t rows, size_t cols) :
+        NMatrix(unsigned long rows, unsigned long cols) :
                 nRows(rows), nCols(cols), data(new T*[rows]){
-            for (size_t i = 0; i < rows; i++)
+            for (unsigned long i = 0; i < rows; i++)
                 data[i] = new T[cols];
         }
         /**
@@ -100,7 +100,7 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          */
         NMatrix(const NMatrix& cloneMe) : nRows(cloneMe.nRows),
                 nCols(cloneMe.nCols), data(new T*[cloneMe.nRows]) {
-            size_t r, c;
+            unsigned long r, c;
             for (r = 0; r < nRows; r++) {
                 data[r] = new T[nCols];
                 for (c = 0; c < nCols; c++)
@@ -111,7 +111,7 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          * Destroys this matrix.
          */
         ~NMatrix() {
-            for (size_t i = 0; i < nRows; i++)
+            for (unsigned long i = 0; i < nRows; i++)
                 delete[] data[i];
             delete[] data;
         }
@@ -122,7 +122,7 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          * @param value the value to assign to each entry.
          */
         void initialise(const T& value) {
-            size_t r, c;
+            unsigned long r, c;
             for (r = 0; r < nRows; r++)
                 for (c = 0; c < nCols; c++)
                     data[r][c] = value;
@@ -150,7 +150,7 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          *
          * @return the number of rows.
          */
-        size_t rows() const {
+        unsigned long rows() const {
             return nRows;
         }
         /**
@@ -158,7 +158,7 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          *
          * @return the number of columns.
          */
-        size_t columns() const {
+        unsigned long columns() const {
             return nCols;
         }
 
@@ -181,7 +181,7 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          * @param column the column of the desired entry.
          * @return a reference to the entry in the given row and column.
          */
-        T& entry(size_t row, size_t column) {
+        T& entry(unsigned long row, unsigned long column) {
             return data[row][column];
         }
         /**
@@ -198,7 +198,7 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          * @param column the column of the desired entry.
          * @return a reference to the entry in the given row and column.
          */
-        const T& entry(size_t row, size_t column) const {
+        const T& entry(unsigned long row, unsigned long column) const {
             return data[row][column];
         }
 
@@ -226,7 +226,7 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
             if (nRows != other.nRows || nCols != other.nCols)
                 return false;
 
-            size_t r, c;
+            unsigned long r, c;
             for (r = 0; r < nRows; ++r)
                 for (c = 0; c < nCols; ++c)
                     if (! (data[r][c] == other.data[r][c]))
@@ -271,7 +271,7 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          * @param out the output stream to which to write.
          */
         void writeMatrix(std::ostream& out) const {
-            size_t r, c;
+            unsigned long r, c;
             for (r = 0; r < nRows; r++) {
                 for (c = 0; c < nCols; c++) {
                     if (c > 0) out << ' ';
@@ -289,9 +289,9 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          * @param first the first row to swap.
          * @param second the second row to swap.
          */
-        void swapRows(size_t first, size_t second) {
+        void swapRows(unsigned long first, unsigned long second) {
             T tmp;
-            for (size_t i = 0; i < nCols; i++) {
+            for (unsigned long i = 0; i < nCols; i++) {
                 tmp = data[first][i];
                 data[first][i] = data[second][i];
                 data[second][i] = tmp;
@@ -305,9 +305,9 @@ class NMatrix : public Output<NMatrix<T> >, public boost::noncopyable {
          * @param first the first column to swap.
          * @param second the second column to swap.
          */
-        void swapColumns(size_t first, size_t second) {
+        void swapColumns(unsigned long first, unsigned long second) {
             T tmp;
-            for (size_t i = 0; i < nRows; i++) {
+            for (unsigned long i = 0; i < nRows; i++) {
                 tmp = data[i][first];
                 data[i][first] = data[i][second];
                 data[i][second] = tmp;
@@ -382,7 +382,7 @@ class NMatrixRing : public NMatrix<T> {
          * @param rows the number of rows in the new matrix.
          * @param cols the number of columns in the new matrix.
          */
-        NMatrixRing(size_t rows, size_t cols) :
+        NMatrixRing(unsigned long rows, unsigned long cols) :
                 NMatrix<T>(rows, cols) {
         }
         /**
@@ -402,7 +402,7 @@ class NMatrixRing : public NMatrix<T> {
          */
         void makeIdentity() {
             this->initialise(zero);
-            for (size_t i = 0; i < this->nRows && i < this->nCols; i++)
+            for (unsigned long i = 0; i < this->nRows && i < this->nCols; i++)
                 this->data[i][i] = one;
         }
 
@@ -422,7 +422,7 @@ class NMatrixRing : public NMatrix<T> {
             if (this->nRows != this->nCols)
                 return false;
 
-            size_t r, c;
+            unsigned long r, c;
             for (r = 0; r < this->nRows; ++r)
                 for (c = 0; c < this->nCols; ++c) {
                     if (r == c && this->data[r][c] != one)
@@ -456,8 +456,8 @@ class NMatrixRing : public NMatrix<T> {
          * @param source the row to add.
          * @param dest the row that will be added to.
          */
-        void addRow(size_t source, size_t dest) {
-            for (size_t i = 0; i < this->nCols; i++)
+        void addRow(unsigned long source, unsigned long dest) {
+            for (unsigned long i = 0; i < this->nCols; i++)
                 this->data[dest][i] += this->data[source][i];
         }
         /**
@@ -475,8 +475,9 @@ class NMatrixRing : public NMatrix<T> {
          * @param copies the number of copies of \a source to add to
          * \a dest.
          */
-        void addRow(size_t source, size_t dest, T copies) {
-            for (size_t i = 0; i < this->nCols; i++)
+        void addRow(unsigned long source, unsigned long dest,
+                T copies) {
+            for (unsigned long i = 0; i < this->nCols; i++)
                 this->data[dest][i] += copies * this->data[source][i];
         }
         /**
@@ -488,8 +489,8 @@ class NMatrixRing : public NMatrix<T> {
          * @param source the columns to add.
          * @param dest the column that will be added to.
          */
-        void addCol(size_t source, size_t dest) {
-            for (size_t i = 0; i < this->nRows; i++)
+        void addCol(unsigned long source, unsigned long dest) {
+            for (unsigned long i = 0; i < this->nRows; i++)
                 this->data[i][dest] += this->data[i][source];
         }
         /**
@@ -507,8 +508,9 @@ class NMatrixRing : public NMatrix<T> {
          * @param copies the number of copies of \a source to add to
          * \a dest.
          */
-        void addCol(size_t source, size_t dest, T copies) {
-            for (size_t i = 0; i < this->nRows; i++)
+        void addCol(unsigned long source, unsigned long dest,
+                T copies) {
+            for (unsigned long i = 0; i < this->nRows; i++)
                 this->data[i][dest] += copies * this->data[i][source];
         }
         /**
@@ -522,8 +524,8 @@ class NMatrixRing : public NMatrix<T> {
          * @param row the row to work with.
          * @param factor the factor by which to multiply the given row.
          */
-        void multRow(size_t row, T factor) {
-            for (size_t i = 0; i < this->nCols; i++)
+        void multRow(unsigned long row, T factor) {
+            for (unsigned long i = 0; i < this->nCols; i++)
                 this->data[row][i] *= factor;
         }
         /**
@@ -537,8 +539,8 @@ class NMatrixRing : public NMatrix<T> {
          * @param column the column to work with.
          * @param factor the factor by which to multiply the given column.
          */
-        void multCol(size_t column, T factor) {
-            for (size_t i = 0; i < this->nRows; i++)
+        void multCol(unsigned long column, T factor) {
+            for (unsigned long i = 0; i < this->nRows; i++)
                 this->data[i][column] *= factor;
         }
 
@@ -568,7 +570,7 @@ class NMatrixRing : public NMatrix<T> {
             std::unique_ptr<NMatrixRing<T> > ans(new NMatrixRing<T>(
                 this->nRows, other.nCols));
 
-            size_t row, col, k;
+            unsigned long row, col, k;
             for (row = 0; row < this->nRows; row++)
                 for (col = 0; col < other.nCols; col++) {
                     ans->data[row][col] = zero;
@@ -605,7 +607,7 @@ class NMatrixRing : public NMatrix<T> {
             std::unique_ptr<MatrixClass> ans(new MatrixClass(
                 this->nRows, other.nCols));
 
-            size_t row, col, k;
+            unsigned long row, col, k;
             for (row = 0; row < this->nRows; row++)
                 for (col = 0; col < other.nCols; col++) {
                     ans->data[row][col] = zero;
@@ -631,7 +633,7 @@ class NMatrixRing : public NMatrix<T> {
          * @return the determinant of this matrix.
          */
         T det() const {
-            size_t n = this->nRows;
+            unsigned long n = this->nRows;
 
             // Just in case...
             if (n != this->nCols || n == 0)
@@ -641,7 +643,7 @@ class NMatrixRing : public NMatrix<T> {
             partial[0] = new T[n * n];
             partial[1] = new T[n * n];
 
-            size_t len, head, curr, prevHead, prevCurr;
+            unsigned long len, head, curr, prevHead, prevCurr;
 
             // Treat the smallest cases of len = 1 separately.
             int layer = 0;
